requirements.txt
# Core ML/Data Science
torch>=2.0.0
pytorch-forecasting>=1.0.0
pytorch-lightning>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Data Collection APIs
entsoe-py>=0.6.0
openmeteo-requests>=1.0.0
requests-cache>=1.0.0
retry-requests>=2.0.0

# Configuration & Environment
python-dotenv>=1.0.0

# API (if deploying the FastAPI service)
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3